Refactor expression evaluator.
Refactoring the expression evaluator into separate types that are
responsible for evaluating access, basic, catalog, control flow, postfix
and primary expressions.
